requirements.txt
Flask
RPi.GPIO
requests